refactor(ssz): move ssz encoder and tests under utils package
