+++ prover-ray/utils/ssz/pubkey_test.go
@@ -1,4 +1,4 @@
-package backend
+package ssz
 
 import (
 	"math/big"
