+++ prover-ray/utils/ssz/ssz_test.go
@@ -1,4 +1,4 @@
-package backend
+package ssz
 
 import (
 	"encoding/json"
@@ -451,20 +451,4 @@ func TestAddSSZLength_RejectsUint32Overflow(t *testing.T) {
 	_, err := addSSZLength("container", maxSSZSerializedBytes, 1)
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "uint32 offset limit")
-}
-
-// TestBuildInputs_MultiBlock_NotImplemented rejects multi-block jobs.
-func TestBuildInputs_MultiBlock_NotImplemented(t *testing.T) {
-	c := &Core{cfg: Config{}}
-	_, err := c.buildInputs(Job{StartBlock: 1, EndBlock: 2, Type: ProofTypeL2Execution})
-	require.ErrorIs(t, err, ErrNotImplemented)
-}
-
-// TestBuildInputs_InvertedRange rejects malformed block ranges.
-func TestBuildInputs_InvertedRange(t *testing.T) {
-	c := &Core{cfg: Config{}}
-	_, err := c.buildInputs(Job{StartBlock: 5, EndBlock: 3, Type: ProofTypeL2Execution})
-	require.Error(t, err)
-	require.NotErrorIs(t, err, ErrNotImplemented)
-	require.Contains(t, err.Error(), "invalid block range")
 }
