+++ prover-ray/utils/ssz/testdata/README.md
@@ -1,10 +1,10 @@
-# backend/testdata
+# utils/ssz/testdata
 
 Golden vectors for `EncodeStatelessInput`, generated by the reference encoder
 `rollup_spec/src/rollup_spec/stateless_input.py::encode_stateless_input_ssz`.
 The Go tests commit both the readable input JSON and the expected framed SSZ
-bytes so `go test ./backend` does not depend on Python, remerkleable, or the
-execution-specs dependency toolchain being installed locally.
+bytes so `go test ./utils/ssz` does not depend on Python, remerkleable, or
+the execution-specs dependency toolchain being installed locally.
 
 | Fixture | Input | Expected output | Length | SHA-256 |
 |---|---|---|---|---|
@@ -73,10 +73,10 @@ for i, payload in enumerate(req['proofRequest']['payloads']):
         'chainConfig': {'chainId': int(cc['chainId']), 'forkName': cc['forkName']},
     }
 
-    with open(f'../prover-ray/backend/testdata/stateless_input_payload{i}.json', 'w') as f:
+    with open(f'../prover-ray/utils/ssz/testdata/stateless_input_payload{i}.json', 'w') as f:
         json.dump(encoder_obj, f, indent=2)
         f.write('\n')
-    with open(f'../prover-ray/backend/testdata/stateless_input_payload{i}.ssz', 'wb') as f:
+    with open(f'../prover-ray/utils/ssz/testdata/stateless_input_payload{i}.ssz', 'wb') as f:
         f.write(encode_stateless_input_ssz(encoder_obj))
 EOF
 ```
@@ -100,11 +100,11 @@ sys.path.insert(0, 'src')
 from rollup_spec.stateless_input import encode_stateless_input_ssz
 
 repo_root = Path('..')
-with open(repo_root / 'prover-ray/backend/testdata/stateless_input_full.json') as f:
+with open(repo_root / 'prover-ray/utils/ssz/testdata/stateless_input_full.json') as f:
     encoder_obj = json.load(f)
 
 sys.stdout.buffer.write(encode_stateless_input_ssz(encoder_obj))
 EOF
 ```
 
-Pipe the output to `prover-ray/backend/testdata/stateless_input_full.ssz`.
+Pipe the output to `prover-ray/utils/ssz/testdata/stateless_input_full.ssz`.
