+++ prover-ray/utils/ssz/testdata/stateless_input_full.ssz
