+++ prover-ray/utils/ssz/testdata/stateless_input_payload0.ssz
