+++ prover-ray/utils/ssz/testdata/stateless_input_payload1.json
